+++ src/Services/Landing/Block/Result/BlockResult.php
@@ -1,0 +1,28 @@
+<?php
+
+/**
+ * This file is part of the bitrix24-php-sdk package.
+ *
+ * © Sally Fancen <vadimsallee@gmail.com>
+ *
+ * For the full copyright and license information, please view the MIT-LICENSE.txt
+ * file that was distributed with this source code.
+ */
+
+declare(strict_types=1);
+
+namespace Bitrix24\SDK\Services\Landing\Block\Result;
+
+use Bitrix24\SDK\Core\Exceptions\BaseException;
+use Bitrix24\SDK\Core\Result\AbstractResult;
+
+class BlockResult extends AbstractResult
+{
+    /**
+     * @throws BaseException
+     */
+    public function getBlock(): BlockItemResult
+    {
+        return new BlockItemResult($this->getCoreResponse()->getResponseData()->getResult());
+    }
+}